fix(Model): remove unncessary function binding for virtuals
from the Object.defineProperty docs:

A function which serves as a getter for the property, or undefined
if there is no getter. When the property is accessed, this function
is called without arguments and with this set to the object through
which the property is accessed (this may not be the object on which
the property is defined due to inheritance). The return value will
be used as the value of the property. Defaults to undefined