fix(Model): do not duplicate Model.config.fieldNames
When a model is inherited and a field overwritten by the child,
`fieldNames` would be duplicated, which would mean that the
`fieldNames` that `Query` uses to cast models before insert would
contain the same field twice. For the case of JSON fields, this
would lead to the field's value being stringified twice.

Closes https://github.com/knorm/postgres/pull/11